+++ lib/cli/buildEbook.js
@@ -25,6 +25,7 @@ module.exports = function(format) {
 
             // Create temporary directory
             var outputFolder = tmp.dirSync().name;
+            options.outputFolder = outputFolder;
 
             var book = getBook(args, kwargs);
             var logger = book.getLogger();
@@ -71,6 +72,7 @@ module.exports = function(format) {
                 logger.info.ok(count + ' file(s) generated');
 
                 logger.debug('cleaning up... ');
+                delete options.outputFolder;
                 return logger.debug.promise(fs.rmDir(outputFolder));
             });
         }
