+++ lib/output/website/createTemplateEngine.js
@@ -91,23 +91,6 @@ function createTemplateEngine(output, currentFile) {
         return JSONUtils.encodePage(page, summary);
     }
 
-    function resolveFileNotWithinRoot(filePath) {
-        if (currentFile.substr(0,2) === '..') {
-            var regex = {
-                "path_sep":    /[\\\/]/,
-                "currentFile": /^((?:\.\.\/)+)([^\.].*)$/
-            };
-
-            var up_count = currentFile.replace(regex.currentFile, '$1').split('/').length - 1;
-            var dn_count = currentFile.replace(regex.currentFile, '$2').split('/').length - 1;
-            var base_path = ('../').repeat(dn_count) + outputFolder.split(regex.path_sep).slice(-up_count).join('/');
-
-            filePath = base_path + '/gitbook/' + filePath;
-            return filePath;
-        }
-        return false;
-    }
-
     return TemplateEngine.create({
         loader: loader,
 
@@ -133,17 +116,11 @@ function createTemplateEngine(output, currentFile) {
              * it also resolve pages
              */
             resolveFile: function(filePath) {
-                var resolvedPath = resolveFileNotWithinRoot(filePath);
-                if (resolvedPath) return resolvedPath;
-
                 filePath = resolveFileToURL(output, filePath);
                 return LocationUtils.relativeForFile(currentFile, filePath);
             },
 
             resolveAsset: function(filePath) {
-                var resolvedPath = resolveFileNotWithinRoot(filePath);
-                if (resolvedPath) return resolvedPath;
-
                 filePath = LocationUtils.toAbsolute(filePath, '', '');
                 filePath = path.join('gitbook', filePath);
                 filePath = LocationUtils.relativeForFile(currentFile, filePath);
