+++ lib/utils/location.js
@@ -97,8 +97,26 @@ function toAbsolute(_href, dir, outdir) {
  * @return {String}
  */
 function relative(dir, file) {
-    var isDirectory = file.slice(-1) === '/';
-    return normalize(path.relative(dir, file)) + (isDirectory? '/': '');
+    if (dir.substr(0,2) === '..') {
+        var outputFolder = require('../cli/options').outputFolder;
+        var currentFile = dir + '/';
+
+        var regex = {
+            "path_sep":    /[\\\/]/,
+            "currentFile": /^((?:\.\.\/)+)([^\.].*)$/
+        };
+
+        var up_count = currentFile.replace(regex.currentFile, '$1').split('/').length - 1;
+        var dn_count = currentFile.replace(regex.currentFile, '$2').split('/').length - 1;
+        var base_path = ('../').repeat(dn_count) + outputFolder.split(regex.path_sep).slice(-up_count).join('/');
+
+        var filePath = base_path + '/' + file;
+        return filePath;
+    }
+    else {
+        var isDirectory = file.slice(-1) === '/';
+        return normalize(path.relative(dir, file)) + (isDirectory? '/': '');
+    }
 }
 
 /**
